Convert JavaScript project to TypeScript with full type safety
