+++ full.ts
@@ -1,0 +1,3 @@
+/*! simple-peer. MIT License. Feross Aboukhadijeh <https://feross.org/opensource> */
+// full.ts is just an alias for index.ts (the complete implementation with stream support)
+export { default, Peer } from './index.js'
