+++ tsconfig.json
@@ -36,6 +36,7 @@
     "node_modules",
     "dist",
     "test",
-    "perf"
+    "perf",
+    "*.js"
   ]
 }
